fix: query mostDangerousPointsWithinRadius (replace $geoNear with $geoWithin)

+++ bbdd2PromocionApp/src/main/java/com/bbdd2promocion/service/impl/AccidentServiceImpl.java
@@ -60,7 +60,7 @@ public class AccidentServiceImpl implements IAccidentService {
     }
 
     @Override
-    public List<LocationCount> getMostDangerousPointsWithinRadius(Double longitude, Double latitude, int radius, int limit){
+    public List<LocationCount> getMostDangerousPointsWithinRadius(Double longitude, Double latitude, Double radius, int limit){
         return this.getAccidentMongoRepository().findMostDangerousPointsWithinRadius(longitude, latitude, radius, limit).getMappedResults();
     }
